Unifying desktop and mobile.
Instead of taking the old approach of having separate code and styling
for mobile and desktop, I have made minor adjustments so that the same
css and html renders appropriately in both mobile and desktop
forms. This greatly simplifies the codebase, even though it restricts
some of my future options for design.